sim-ln/refactor: move parsing into its own module
